refactor(clients): extract TableColumnSelect
