+++ src/components.d.ts
@@ -21,6 +21,7 @@ declare module 'vue' {
     CertBundleInfo: typeof import('./components/TLSConfig/CertBundleInfo.vue')['default']
     CertFileInput: typeof import('./components/TLSConfig/CertFileInput.vue')['default']
     CheckIcon: typeof import('./components/CheckIcon.vue')['default']
+    ClientFieldSelect: typeof import('./components/ClientFieldSelect.vue')['default']
     CodeView: typeof import('./components/CodeView.vue')['default']
     CommonItemStatus: typeof import('./components/CommonItemStatus.vue')['default']
     CommonMiniPagination: typeof import('./components/CommonMiniPagination.vue')['default']
@@ -85,6 +86,7 @@ declare module 'vue' {
     StatsContent: typeof import('./components/StatsContent.vue')['default']
     StatusDetailsOfEachNode: typeof import('./components/StatusDetailsOfEachNode.vue')['default']
     TableButton: typeof import('./components/Buttons/TableButton.vue')['default']
+    TableColumnSelect: typeof import('./components/TableColumnSelect.vue')['default']
     TextareaWithUploader: typeof import('./components/TextareaWithUploader.vue')['default']
     TextEasyCopy: typeof import('./components/TextEasyCopy.vue')['default']
     TimeInputWithUnitSelect: typeof import('./components/TimeInputWithUnitSelect.vue')['default']
